Fix hue interpolation
Since we are in a looped space, it makes sense to adjust the interpolation to take shorter paths through the space. A simple lerp may take the long path through some funky colors, even when the two colors are quite close. For example, blue -> red would go through green. We don't want that.