Commit progress
- Drop elm-animation. The biggest problem with it is the hard-coded
spring-like re-targeting. I need a lot more control when it comes to
animation, so I'll end up rolling my own.
- Remove all animation code.
- Update the time on each animation frame. It doesn't make sense to have
to loops: one for the time and one for animations. The worst thing is
that we can end up in a situation where the seconds arm skips 2 seconds
ahead because of the difference in update cadence.
- Move the clock view functions into the `Clock` module.
- Add sub-units to ticks. We can now compute intermediate angles for
very accurate continuous motion.
- Add option to create a `none` color — basically transparent.
- Fix maximum corner radius calculation.